+++ cypress/e2e/testUiAndFunction/startPage/US_01.03_Header_elements.cy.js
@@ -1,8 +1,8 @@
 /// <reference types="Cypress" />
 
-import {StartPage} from "../../../../pageObjects/StartPage.js";
-import {LoginPopup} from "../../../../pageObjects/StartPage.js";
-import Header from "../../../../pageObjects/Header.js";
+import {StartPage} from "../../../pageObjects/StartPage.js";
+import {LoginPopup} from "../../../pageObjects/StartPage.js";
+import Header from "../../../pageObjects/Header.js";
 
 const startPage = new StartPage();
 const loginPopup = new LoginPopup();
