+++ cypress/e2e/testUiAndFunction/startPage/US_01.05_LoginByEmailTab_UI.cy.js
@@ -1,7 +1,7 @@
 /// <reference types="Cypress" />
 
-import {StartPage} from "../../../../pageObjects/StartPage.js";
-import {LoginPopup} from "../../../../pageObjects/StartPage.js";
+import {StartPage} from "../../../pageObjects/StartPage.js";
+import {LoginPopup} from "../../../pageObjects/StartPage.js";
 
 const startPage = new StartPage();
 const loginPopup = new LoginPopup();
