+++ cypress/e2e/testUiAndFunction/startPage/US_01.06_LoginByEmailTub.cy.js
@@ -1,6 +1,6 @@
 /// <reference types="Cypress" />
 
-import CreateBookingPage from "../../../../pageObjects/CreateBookingPage";
+import CreateBookingPage from "../../../pageObjects/CreateBookingPage";
 
 const createBookingPage = new CreateBookingPage();
